fix: send 200 on user id already existing
