--- tests/food.test.ts
@@ -1,3 +1,0 @@
-import { createTestWithUser } from "./fixtures";
-
-const test = createTestWithUser("food");
